+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Fixed some minor bugs on enterprise plain text captured password process
  - Tested compatibility with Wifislax 64-2.x
  - Fixed dependency problem to avoid use of xxd
+ - Routing rules adapted to work with nftables
  
 ### 9.01
  - Improvments on options environment var system